+++ memberaudit/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = "memberaudit.apps.MemberAuditConfig"
 
-__version__ = "1.3.2"
+__version__ = "1.3.3"
 __title__ = "Member Audit"